[ES-DE Hotfix] (#1013)
* If migration is not done first, .emulationstation directory is moved inside of the newly created ESDE folder, moved migration to ToolScripts

+++ functions/ToolScripts/emuDeckESDE.sh
@@ -34,7 +34,7 @@ ESDE_cleanup(){
 
 ESDE_migration(){
 
-	if [ -d "$ESDE_oldConfigDirectory" ] && [ ! -L "$ESDE_oldConfigDirectory" ]; then
+	if [ -d "$ESDE_oldConfigDirectory" ] && [ ! -L "$ESDE_oldConfigDirectory" ] && [ ! -d "$ESDE_newConfigDirectory" ]; then
 		mv "$ESDE_oldConfigDirectory" "$ESDE_newConfigDirectory"
 		ln -s  "$ESDE_newConfigDirectory" "$ESDE_oldConfigDirectory"
 		echo "EmulationStation-DE config directory successfully migrated and linked."
@@ -81,6 +81,9 @@ ESDE_install(){
 	ESDE_migration
 	
 	ESDE_SetAppImageURLS
+	ESDE_migration
+
+
 	local showProgress="$1"
 	echo $ESDE_releaseURL
 	if [[ $ESDE_releaseURL = "https://gitlab.com/es-de/emulationstation-de/-/package_files/"* ]]; then
@@ -100,6 +103,8 @@ ESDE_install(){
 ESDE_init(){
 	setMSG "Setting up $ESDE_toolName"
 
+	ESDE_migration
+
 	mkdir -p "$ESDE_newConfigDirectory/custom_systems/"
 
 	rsync -avhp --mkpath "$EMUDECKGIT/configs/emulationstation/es_settings.xml" "$(dirname "$es_settingsFile")" --backup --suffix=.bak
@@ -115,7 +120,6 @@ ESDE_init(){
 	#ESDE_addSteamInputProfile
 	ESDE_symlinkGamelists
 	ESDE_migrateEpicNoir
-	ESDE_migration
 
 	if [ "$system" == "chimeraos" ] || [ "$system" == "ChimeraOS" ]; then
 			ESDE_chimeraOS
@@ -141,6 +145,8 @@ ESDE_resetConfig(){
 ESDE_update(){
 	setMSG "Setting up $ESDE_toolName"
 
+	ESDE_migration
+
 	mkdir -p "$ESDE_newConfigDirectory/custom_systems/"
 
 	#update es_settings.xml
@@ -154,7 +160,6 @@ ESDE_update(){
 	ESDE_migrateDownloadedMedia
 	#ESDE_addSteamInputProfile
 	ESDE_symlinkGamelists
-	ESDE_migration
 }
 
 ESDE_addCustomSystems(){
